# Improve `isIdentifierName` performance