--- a/packages/babel-helper-validator-identifier/.npmignore
+++ b/packages/babel-helper-validator-identifier/.npmignore
@@ -1,3 +1,4 @@
+benchmark
 src
 test
 *.log