--- a/packages/babel-helper-validator-identifier/benchmark/identifier.bench.mjs
+++ b/packages/babel-helper-validator-identifier/benchmark/identifier.bench.mjs
@@ -1,0 +1,3 @@
+import "./isIdentifierChar.bench.mjs";
+import "./isIdentifierStart.bench.mjs";
+import "./isIdentifierName.bench.mjs";