--- a/packages/babel-helper-validator-identifier/benchmark/isIdentifierName.bench.mjs
+++ b/packages/babel-helper-validator-identifier/benchmark/isIdentifierName.bench.mjs
@@ -5,34 +5,26 @@ import { report } from "./util.mjs";
 
 const suite = new Benchmark.Suite();
 
-suite.add("baseline#isIdentifierName on 2 short ASCII words", () => {
-  baseline.isIdentifierName("aforementioned");
-  baseline.isIdentifierName("zap cannon");
-});
+function benchCases(implementation, name) {
+  suite.add(name + "#isIdentifierName on 2 short ASCII words", () => {
+    implementation.isIdentifierName("aforementioned");
+    implementation.isIdentifierName("zap cannon");
+  });
 
-suite.add("baseline#isIdentifierName on 1 long ASCII words", () => {
-  baseline.isIdentifierName("Pneumonoultramicroscopicsilicovolcanoconiosis");
-});
+  suite.add(name + "isIdentifierName on 1 long ASCII words", () => {
+    implementation.isIdentifierName(
+      "Pneumonoultramicroscopicsilicovolcanoconiosis"
+    );
+  });
 
-suite.add("baseline#isIdentifierName on 3 non-ASCII words", () => {
-  baseline.isIdentifierName("مذكور أعلاه");
-  baseline.isIdentifierName("cañón de zap");
-  baseline.isIdentifierName("𠡦𠧋𡆠囝〇𠁈𢘑𤯔𠀑埊");
-});
+  suite.add(name + "isIdentifierName on 3 non-ASCII words", () => {
+    implementation.isIdentifierName("مذكور أعلاه");
+    implementation.isIdentifierName("cañón de zap");
+    implementation.isIdentifierName("𠡦𠧋𡆠囝〇𠁈𢘑𤯔𠀑埊");
+  });
+}
 
-suite.add("current#isIdentifierName on 2 short ASCII words", () => {
-  current.isIdentifierName("aforementioned");
-  current.isIdentifierName("zap cannon");
-});
-
-suite.add("current#isIdentifierName on 1 long ASCII words", () => {
-  current.isIdentifierName("Pneumonoultramicroscopicsilicovolcanoconiosis");
-});
-
-suite.add("current#isIdentifierName on 3 non-ASCII words", () => {
-  current.isIdentifierName("مذكور أعلاه");
-  current.isIdentifierName("cañón de zap");
-  current.isIdentifierName("𠡦𠧋𡆠囝〇𠁈𢘑𤯔𠀑埊");
-});
+benchCases(baseline, "baseline");
+benchCases(current, "current");
 
 suite.on("cycle", report).run();