--- a/packages/babel-helper-validator-identifier/package.json
+++ b/packages/babel-helper-validator-identifier/package.json
@@ -14,7 +14,9 @@
   "main": "./lib/index.js",
   "exports": "./lib/index.js",
   "devDependencies": {
-    "charcodes": "^0.2.0",
-    "unicode-13.0.0": "^0.8.0"
+    "@babel/helper-validator-identifier-baseline": "npm:@babel/helper-validator-identifier@7.10.4",
+    "@unicode/unicode-13.0.0": "^1.0.6",
+    "benchmark": "^2.1.4",
+    "charcodes": "^0.2.0"
   }
 }