--- a/packages/babel-helper-validator-identifier/scripts/generate-identifier-regex.js
+++ b/packages/babel-helper-validator-identifier/scripts/generate-identifier-regex.js
@@ -4,14 +4,14 @@
 // https://tc39.github.io/ecma262/#sec-conformance
 const version = "13.0.0";
 
-const start = require("unicode-" +
+const start = require("@unicode/unicode-" +
   version +
   "/Binary_Property/ID_Start/code-points.js").filter(function (ch) {
   return ch > 0x7f;
 });
 let last = -1;
 const cont = [0x200c, 0x200d].concat(
-  require("unicode-" +
+  require("@unicode/unicode-" +
     version +
     "/Binary_Property/ID_Continue/code-points.js").filter(function (ch) {
     return ch > 0x7f && search(start, ch, last + 1) == -1;