<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Any Dependency Changes?  | Added benchmark to `devDependencies`
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
Results from my local machine
```sh
$ node packages/babel-helper-validator-identifier/benchmark/isIdentifierName.bench.mjs
```

```
baseline#isIdentifierName on 2 short ASCII words: 2819979 ops/sec ±1.14% (0ms)
baselineisIdentifierName on 1 long ASCII words: 1486317 ops/sec ±0.4% (0.001ms)
baselineisIdentifierName on 3 non-ASCII words: 250140 ops/sec ±0.38% (0.004ms)
current#isIdentifierName on 2 short ASCII words: 7775830 ops/sec ±0.38% (0ms)
currentisIdentifierName on 1 long ASCII words: 4291478 ops/sec ±0.33% (0ms)
currentisIdentifierName on 3 non-ASCII words: 245762 ops/sec ±0.53% (0.004ms)
```

Initially meant to test per-package benchmark setup, this PR sees up to 3x improvement for ASCII words. The function is not used by `@babel/parser` internally, instead it is exported via `@babel/types` and used in systemjs and JSX transforms.

This PR also adds more test cases to `isIdentifier*`.