@bmeurer 2021-04-28
I'm pretty sure the major performance difference is due to not using `Array.from`. Beyond that, I think it'd be good to file a feature request for V8 to fix the performance cliff with `codePointAt` in TurboFan.