@jridgewell 2021-05-06
@JLHwung did you open an v8 issue about `codePointAt`'s perf cliff?