@JLHwung 2021-05-06
For reference I open a V8 issue: https://bugs.chromium.org/p/v8/issues/detail?id=11743